NFC: Add a publish-function-api only GHA. Restrict publishing to only recent versions.

+++ tool/cicd/actions/scripts/publish_fusebit_ops_cli.sh
@@ -13,6 +13,15 @@ VERSION=${VERSION_FUSEBIT_OPS_CLI:=`jq -r '.version' ./cli/fusebit-ops-cli/packa
 AWS_S3_OPTS="--acl public-read --cache-control max-age=300"
 MAJOR_MINOR=`echo ${VERSION} | sed 's/\([0-9]*\)\.\([0-9]*\)\.\([0-9]*\)/\1.\2/'`
 
+# -- Is this the HEAD of this artifact?
+VER_WART=ops-cli
+git tag --points-at HEAD | grep ${VER_WART}-${VERSION} > /dev/null;
+TAG_TEST=$?;
+if [ ${TAG_TEST} -ne 0 ]; then
+  echoerr "Not publishing ${VERSION} - HEAD is not tagged ${VER_WART}-${VERSION}";
+  exit 0;
+fi
+
 # -- Script --
 echoerr "Creating fusebit-ops-cli package: ${VERSION} at ${MAJOR_MINOR}"
 rm -rf cli/fusebit-ops-cli/package
